방어력 계산식완성

+++ Code Practice 2D/Assets/Script/Player/PlayerInfo.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 public class PlayerInfo : MonoBehaviour
 {
     public Animator anim;
-    public int HP = 100;
-    public int att = 10;
-    public int def = 20;
-    public int mana = 100;
+    public float HP = 100;
+    public float att = 10;
+    public float def = 20f;
+    public float mana = 100;
     public float speed = 3f;
     public float jumpSpeed = 5f;
     public float dashSpeed = 10f;
